1.增加多种gm命令的支持/2.修改机器人用于测试GM

+++ cat-server/src/main/java/com/cat/server/game/module/playermail/domain/PlayerMailDomain.java
@@ -6,11 +6,10 @@ import org.slf4j.LoggerFactory;
 
 
 /**
-* PlayerMailDomain
+* PlayerMailDomain, 
 * @author Jeremy
 */
-
-public class PlayerMailDomain extends AbstractModuleMultiDomain<Long, Integer, PlayerMail> {
+public class PlayerMailDomain extends AbstractModuleMultiDomain<Long, Long, PlayerMail> {
 
 	private static final Logger log = LoggerFactory.getLogger(PlayerMailDomain.class);
